Module voca_rs::case

Converts the subject to a selected case.

Functions

camel_case
Converts the subject to camel case.

capitalize
Converts the first character of subject to upper case. If restToLower is true, convert the rest of subject to lower case.

decapitalize
Converts the first character of subject to lower case. If restToLower is true, convert the rest of subject to lower case.

kebab_case
Converts the subject to kebab case.

lower_case
Converts the subject to lower case.

lower_first
Converts the first character of the subject to lower case.

pascal_case
Converts the subject to pascal case.

shouty_kebab_case
Converts the subject to SHOUTY kebab case.

shouty_snake_case
Converts the subject to SHOUTY snake case.

snake_case
Converts the subject to snake case.

swap_case
Converts the uppercase alpha characters of subject to lowercase and lowercase characters to uppercase.

title_case
Converts the subject to title case.

train_case
Converts the subject to train case.

upper_case
Converts the subject to upper case.

upper_first
Converts the first character of the subject to upper case.