Module voca_rs::chop

Extracts a character(s) from subject.

Functions

char_at
Access a character from subject at specified position.

code_point_at
Get the Unicode code point value of the character at position. NOTE: Unicode escape must not be a surrogate

first
Extracts the first length characters from subject.

foreign_key
Converts the subject to a foreign_key.

grapheme_at
Get a grapheme from subject at specified position.

last
Extracts the last length characters from subject.

max
Returns the max character from the subject by its code point. NOTE: Unicode escape must not be a surrogate

min
Returns the min character from the subject by its code point. NOTE: Unicode escape must not be a surrogate

prune
Truncates subject to a new length and does not break the words. Guarantees that the truncated string is no longer than length.

slice
Extracts from subject a string from start position up to end position. The character at end position is not included.

substr
Extracts from subject a string from start position a number of length characters.

substring
Extracts from subject a string from start position up to end position. The character at end position is not included.

truncate
Truncates subject to a new length.