Module voca_rs::count

Counts the characters in subject.

Functions

count
Counts the characters in subject.

count_graphemes
Counts the graphemes in subject taking care of surrogate pairs and combining marks.

count_substrings
Counts the number of substring appearances in subject.

count_unique_words
Counting occurrences of unique words in subject. This function respects unicode.

count_where
Counts the characters in subject for which predicate returns true. This function respects unicode.

count_words
Counts the number of words in subject.