Module voca_rs::query

Checks a subject against a query.

Functions

ends_with
Checks whether subject ends with end.

includes
Checks whether subject includes search starting from position.

is_alpha
Checks whether subject contains only alpha characters.

is_alphadigit
Checks whether subject contains contains only alpha and digit characters.

is_blank
Checks whether subject is empty or contains only whitespaces.

is_digit
Checks whether subject contains only digit characters.

is_empty
Checks whether subject is empty.

is_lowercase
Checks whether subject has only lower case characters.

is_numeric
Checks whether subject is numeric.

is_uppercase
Checks whether subject has only upper case characters.

starts_with
Checks whether subject starts with start.