Module voca_rs::utils

Utility functions and properties.

Constants

ASCII_LETTERS
The concatenation of the ascii_lowercase and ascii_uppercase constants described below. This value is not locale-dependent.

ASCII_LOWERCASE
The lowercase letters "abcdefghijklmnopqrstuvwxyz". This value is not locale-dependent and will not change.

ASCII_UPPERCASE
The uppercase letters "ABCDEFGHIJKLMNOPQRSTUVWXYZ". This value is not locale-dependent and will not change.

DIGITS
The string "0123456789".

HEXDIGITS
The string "0123456789abcdefABCDEF".

OCTDIGITS
The string "01234567".

PRINTABLE
String of ASCII characters which are considered printable. This is a combination of digits, ascii_letters, punctuation, and whitespace.

PUNCTUATION
The string !"#$%&'()*+,-./:;<=>?@[\]^_{|}~`.

VERSION
A property that contains the library semantic version number.

WHITESPACE
The string " \t\n\r".

Functions

unidecode
Takes any Unicode string and returns an ASCII transliteration of that string. https://github.com/chowdhurya/rust-unidecode/blob/master/src/lib.rs

unidecode_char
Takes a single Unicode character and returns an ASCII transliteration. https://github.com/chowdhurya/rust-unidecode/blob/master/src/lib.rs