Module voca_rs::query

Checks a subject against a query.

Functions

ends_with
Checks whether subject ends with end.

includes
Checks whether subject includes search starting from position.

is_alpha
Checks whether subject contains only alpha characters.

is_alphadigit
Checks whether subject contains contains only alpha and digit characters.

is_blank
Checks whether subject is empty or contains only whitespaces.

is_camel_case
Checks whether subject is camelCased.

is_capitalize
Checks whether subject is capitalized and the rest of subject is converted to lower case.

is_decapitalize
Checks whether subject is decapitalized and the rest of subject is converted to lower case.

is_digit
Checks whether subject contains only digit characters.

is_empty
Checks whether subject is empty.

is_foreign_key
Checks whether subject is is a foreign_key.

is_kebab_case
Checks whether subject is kebab-cased.

is_lower_first
Checks whether subject has the first character in lower case.

is_lowercase
Checks whether subject has only lower case characters.

is_numeric
Checks whether subject is numeric.

is_pascal_case
Checks whether subject is PascalCased.

is_shouty_kebab_case
Checks whether subject is SHOUTY-KEBAB-CASED.

is_shouty_snake_case
Checks whether subject is SHOUTY_SNAKE_CASED.

is_snake_case
Checks whether subject is snake_cased.

is_title
Checks whether subject is a titlecased string and there is at least one character.

is_train_case
Checks whether subject is Train-Cased.

is_upper_first
Checks whether subject has the first character in upper case.

is_uppercase
Checks whether subject has only upper case characters.

matches
Checks whether subject matches the regular expression pattern. NOTE: Executes regular expressions only on valid UTF-8 while exposing match locations as byte indices into the search string (see case #4).

query
Checks whether subject contains all characters from search starting from position. Respects an order of characters.

starts_with
Checks whether subject starts with start.